in progress

+++ test/testkit/k8s/objects/virtual_service.go
@@ -95,7 +95,7 @@ func (s *VirtualService) K8sObject() *istionetworkingclientv1.VirtualService {
 								Value: s.faultAbortPercentage,
 							},
 								ErrorType: &istionetworkingv1.HTTPFaultInjection_Abort_HttpStatus{
-									HttpStatus: int32(s.faultAbortHttpStatus),
+									HttpStatus: s.faultAbortHttpStatus,
 								},
 							}
 						}(),
